FreeCRMTestAutomation_by_NaveenAutomationLabs_Part_5/src/main/java/com/crm/qa/pages/TasksPage.java
package com.crm.qa.pages;

import org.openqa.selenium.WebElement;
import org.openqa.selenium.support.FindBy;
import org.openqa.selenium.support.PageFactory;

import com.crm.qa.base.TestBase;
import com.crm.qa.util.TestUtilityOldKeep;

public class TasksPage extends TestBase {

	// Page Factory - Object Repository
	
	// xpath built using ChroPath, which is Google Chrome extension
	// WebElement tasksPageLabel = driver.findElement(By.xpath("//body/div[@id='ui']/div[1]/div[2]/div[2]/div[1]/div[1]/div[1]")
	@FindBy(xpath="//body/div[@id='ui']/div[1]/div[2]/div[2]/div[1]/div[1]/div[1]") WebElement tasksPageLabel;
		
	// Initializing the Page objects
	public TasksPage() {	
		//PageFactory.initElements(driver,TasksPage.class);
		// "this" is same as "TasksPage.class"
		PageFactory.initElements(driver,this);
	}
		
	// Actions
		
	public boolean verifyLandedOnTasksPage() {
		TestUtilityOldKeep.highLightElement(driver, tasksPageLabel);
		return tasksPageLabel.isDisplayed();
	}
}
